Move 'WFSFilterParser' that is only used in analysis to analysis package
